@@ -11,9 +11,9 @@ describe "Japanese (ja) translations" do
   end
 
   it do
-    I18n.t("backordering_is_allowed", :not => "").should == "取り寄せ可"
-    I18n.t("backordering_is_allowed", :not => I18n.t("not")).should == "取り寄せ不可"
-    I18n.t("products_with_zero_inventory_display", :not => "").should == "在庫なしの商品が表示されます"
-    I18n.t("products_with_zero_inventory_display", :not => I18n.t("not")).should == "在庫なしの商品は表示されません"
+    I18n.t("backordering_is_allowed", not: "").should == "取り寄せ可"
+    I18n.t("backordering_is_allowed", not: I18n.t("not")).should == "取り寄せ不可"
+    I18n.t("products_with_zero_inventory_display", not: "").should == "在庫なしの商品が表示されます"
+    I18n.t("products_with_zero_inventory_display", not: I18n.t("not")).should == "在庫なしの商品は表示されません"
   end
 end
