@@ -2,7 +2,7 @@ RSpec::Matchers.define :be_a_thorough_translation_of do |default_locale_filepath
   match do |filepath|
     locale_file = I18nSpec::LocaleFile.new(filepath)
     default_locale = I18nSpec::LocaleFile.new(default_locale_filepath)
-    
+
     @misses = default_locale.flattened_translations.select do |key, value|
       !@keys.include?(key) &&
         locale_file.flattened_translations[key] != "" &&
@@ -10,7 +10,7 @@ RSpec::Matchers.define :be_a_thorough_translation_of do |default_locale_filepath
     end
     @misses.empty?
   end
-    
+
   chain :except do |keys|
     @keys = keys
   end
